move EncodeSpan to the encoder
