+++ exporter/elasticsearchexporter/internal/mapping/default.go
@@ -4,10 +4,21 @@
 package mapping // import "github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/mapping"
 
 import (
+	"encoding/json"
+	"time"
+
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/plog"
+	"go.opentelemetry.io/collector/pdata/ptrace"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/objmodel"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/traceutil"
+)
+
+const (
+	traceIDField   = "traceID"
+	spanIDField    = "spanID"
+	attributeField = "attribute"
 )
 
 // DefaultEncoder is an encoder that handles the `none` and `raw` encoding modes.
@@ -15,7 +26,7 @@ type DefaultEncoder struct {
 	Mode
 }
 
-func (e DefaultEncoder) EncodeLog(resource pcommon.Resource, record plog.LogRecord, scopeLogs plog.ScopeLogs) objmodel.Document {
+func (e DefaultEncoder) EncodeLog(resource pcommon.Resource, scopeLogs plog.ScopeLogs, record plog.LogRecord) objmodel.Document {
 	var document objmodel.Document
 
 	docTimeStamp := record.Timestamp()
@@ -36,12 +47,41 @@ func (e DefaultEncoder) EncodeLog(resource pcommon.Resource, record plog.LogReco
 	return document
 }
 
+func (e DefaultEncoder) EncodeSpan(resourceSpans ptrace.ResourceSpans, scopeSpans ptrace.ScopeSpans, span ptrace.Span) objmodel.Document {
+	var document objmodel.Document
+
+	document.AddTimestamp("@timestamp", span.StartTimestamp()) // We use @timestamp in order to ensure that we can index if the default data stream logs template is used.
+	document.AddTimestamp("EndTimestamp", span.EndTimestamp())
+	document.AddTraceID("TraceId", span.TraceID())
+	document.AddSpanID("SpanId", span.SpanID())
+	document.AddSpanID("ParentSpanId", span.ParentSpanID())
+	document.AddString("Name", span.Name())
+	document.AddString("Kind", traceutil.SpanKindStr(span.Kind()))
+	document.AddInt("TraceStatus", int64(span.Status().Code()))
+	document.AddString("TraceStatusDescription", span.Status().Message())
+	document.AddString("Link", spanLinksToString(span.Links()))
+	encodeAttributes(&document, e.Mode, span.Attributes())
+	document.AddAttributes("Resource", resourceSpans.Resource().Attributes())
+	encodeEvents(&document, e.Mode, span.Events())
+	document.AddInt("Duration", durationAsMicroseconds(span.StartTimestamp().AsTime(), span.EndTimestamp().AsTime())) // unit is microseconds
+	document.AddAttributes("Scope", scopeToAttributes(scopeSpans.Scope()))
+	return document
+}
+
 func encodeAttributes(document *objmodel.Document, m Mode, attributes pcommon.Map) {
 	key := "Attributes"
 	if m == ModeRaw {
 		key = ""
 	}
 	document.AddAttributes(key, attributes)
+}
+
+func encodeEvents(document *objmodel.Document, m Mode, events ptrace.SpanEventSlice) {
+	key := "Events"
+	if m == ModeRaw {
+		key = ""
+	}
+	document.AddEvents(key, events)
 }
 
 func scopeToAttributes(scope pcommon.InstrumentationScope) pcommon.Map {
@@ -52,4 +92,24 @@ func scopeToAttributes(scope pcommon.InstrumentationScope) pcommon.Map {
 		attrs.PutStr(k, v.(string))
 	}
 	return attrs
+}
+
+func spanLinksToString(spanLinkSlice ptrace.SpanLinkSlice) string {
+	linkArray := make([]map[string]any, 0, spanLinkSlice.Len())
+	for i := 0; i < spanLinkSlice.Len(); i++ {
+		spanLink := spanLinkSlice.At(i)
+		link := map[string]any{}
+		link[spanIDField] = traceutil.SpanIDToHexOrEmptyString(spanLink.SpanID())
+		link[traceIDField] = traceutil.TraceIDToHexOrEmptyString(spanLink.TraceID())
+		link[attributeField] = spanLink.Attributes().AsRaw()
+		linkArray = append(linkArray, link)
+	}
+	linkArrayBytes, _ := json.Marshal(&linkArray)
+	return string(linkArrayBytes)
+}
+
+// durationAsMicroseconds calculate span duration through end - start nanoseconds and converts time.Time to microseconds,
+// which is the format the Duration field is stored in the Span.
+func durationAsMicroseconds(start, end time.Time) int64 {
+	return (end.UnixNano() - start.UnixNano()) / 1000
 }
