+++ exporter/elasticsearchexporter/model_test.go
@@ -200,113 +200,6 @@ func mockResourceLogs() plog.ResourceLogs {
 	return rl
 }
 
-func TestEncodeAttributes(t *testing.T) {
-	t.Parallel()
-
-	attributes := pcommon.NewMap()
-	err := attributes.FromRaw(map[string]any{
-		"s": "baz",
-		"o": map[string]any{
-			"sub_i": 19,
-		},
-	})
-	require.NoError(t, err)
-
-	tests := map[string]struct {
-		mappingMode mapping.Mode
-		want        func() objmodel.Document
-	}{
-		"raw": {
-			mappingMode: mapping.ModeRaw,
-			want: func() objmodel.Document {
-				return objmodel.DocumentFromAttributes(attributes)
-			},
-		},
-		"none": {
-			mappingMode: mapping.ModeNone,
-			want: func() objmodel.Document {
-				doc := objmodel.Document{}
-				doc.AddAttributes("Attributes", attributes)
-				return doc
-			},
-		},
-		"ecs": {
-			mappingMode: mapping.ModeECS,
-			want: func() objmodel.Document {
-				doc := objmodel.Document{}
-				doc.AddAttributes("Attributes", attributes)
-				return doc
-			},
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			m := encodeModel{
-				mode: test.mappingMode,
-			}
-
-			doc := objmodel.Document{}
-			m.encodeAttributes(&doc, attributes)
-			require.Equal(t, test.want(), doc)
-		})
-	}
-}
-
-func TestEncodeEvents(t *testing.T) {
-	t.Parallel()
-
-	events := ptrace.NewSpanEventSlice()
-	events.EnsureCapacity(4)
-	for i := 0; i < 4; i++ {
-		event := events.AppendEmpty()
-		event.SetTimestamp(pcommon.NewTimestampFromTime(time.Now().Add(time.Duration(i) * time.Minute)))
-		event.SetName(fmt.Sprintf("event_%d", i))
-	}
-
-	tests := map[string]struct {
-		mappingMode mapping.Mode
-		want        func() objmodel.Document
-	}{
-		"raw": {
-			mappingMode: mapping.ModeRaw,
-			want: func() objmodel.Document {
-				doc := objmodel.Document{}
-				doc.AddEvents("", events)
-				return doc
-			},
-		},
-		"none": {
-			mappingMode: mapping.ModeNone,
-			want: func() objmodel.Document {
-				doc := objmodel.Document{}
-				doc.AddEvents("Events", events)
-				return doc
-			},
-		},
-		"ecs": {
-			mappingMode: mapping.ModeECS,
-			want: func() objmodel.Document {
-				doc := objmodel.Document{}
-				doc.AddEvents("Events", events)
-				return doc
-			},
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			m := encodeModel{
-				mode: test.mappingMode,
-			}
-
-			doc := objmodel.Document{}
-			m.encodeEvents(&doc, events)
-			require.Equal(t, test.want(), doc)
-		})
-	}
-}
-
 func TestEncodeLogECSModeDuplication(t *testing.T) {
 	resource := pcommon.NewResource()
 	err := resource.Attributes().FromRaw(map[string]any{
